libmei: reset errno before operations
Some operations not resetting errno variable if succeeded.
If errno is important to us - reset it before operation.

From: man 3 errno
For some system calls and library functions (e.g., getpriority(2)),
-1 is a valid return on success. In such cases, a successful return
can be distinguished from an error return by setting errno to zero
before the call, and then, if the call returns a status that
indicates that an error may have occurred, checking to see if errno
has a nonzero value.

Change-Id: I9047ab5465fc964bb62270e83214b39d9c5329a5
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>